+++ Data/Models/Product.cs
@@ -11,8 +11,8 @@ namespace Data.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
-        public decimal Discount { get; set; }
+        public double Price { get; set; }
+        public double Discount { get; set; }
         [ForeignKey("CategoryId")]
         public int CategoryId { get; set; }
         public string MainImage { get; set; }
